Update requirements.txt (#2800)
Hi! I'm the author and one of the maintainers of [PyNomaly](https://github.com/vc1492a/PyNomaly). 

This proposed change fixes the [PyNomaly](https://github.com/vc1492a/PyNomaly) version used by deepchecks to to `0.3.4` until further notice to ensure continued reliability and functionality of deepchecks while we undergo some refactoring and API changes that will alter the manner in which users of our software pass data and fit the Local Outlier Probabilities (LoOP) algorithm to their data. 

Following our refactoring and release of PyNomaly `1.0.0`, we will issue a subsequent pull request to pull the latest version of PyNomaly and update the deepchecks code base to leverage the new API.